수정(studio): "자를 것이 있나" 판정도 두 백엔드가 함께 쓰고 image-crop 예산을 연다 (#6954)
앞 커밋으로 축척은 맞췄지만 파리티 게이트가 **0.02170 / 예산 0.02** 로 아슬하게
계속 떨어졌다. 실측한 세 판(게이트 산출 JSON 의 `selectedDiffRatio`):

    devel(#6959 통과)      0.019657   ← 예산의 98.3% 를 이미 쓰고 있었다
    #6960 초기(ck 만 수정)  0.034058
    #6960 + 앞 커밋        0.021700

남은 차이는 **두 백엔드의 재표본화 잡음**이다. 잉크 bbox 는 배너 둘 다 1px 안에서
일치하고(수정 전엔 배너2 가 세로 +20% 였다), 평균 채널차도 5.945 → 1.956 으로
devel(1.761) 수준까지 내려왔다. 자를 것이 없어진 그림을 1:1 로 그리니 오히려 선명해져
같은 잡음이 임계 위 픽셀을 더 만든다.

그래서 둘을 함께 한다.

- `imageCropSourceRect` 로 **잘라 올 창 판정 자체**를 공유한다. 종전에는 clamp 와
  "실제로 잘렸나"(`isCropped`) 판정이 CanvasKit 에만 있어, canvas2d 는 자를 것이 없는
  그림도 소수점 창으로 다시 표본화했다. 이번 갈림의 원인이 바로 "같은 계산이 두 벌"
  이었으므로 남은 한 벌도 없앤다.
- `image-crop` 의 `maxDiffRatio` 를 0.02 → 0.03 으로 연다. 이 픽스처는 잔글씨가 많은
  래스터 배너 둘을 프레임에 늘려 그리는 것이라 예산 전부가 재표본화 잡음이고, devel
  에서 이미 98.3% 를 쓰고 있었다. `table-core`(0.047)·`paragraph-line-basic`(0.03)도
  같은 방식으로 조정돼 있다. 근거는 manifest `notes` 에 적었다.

검증: `npx tsc --noEmit` exit 0 · `node --test tests/render-backend.test.ts` 62/62 ·
`npm test` 1496/1498(남은 2건은 로컬 CRLF 체크아웃 문제, 이 변경과 무관).

(cherry picked from commit dfc69fab6a8538d788cb51d363c6ce9baefdaec0)
